requirements.txt
fastapi
uvicorn
celery
redis
boto3
python-dotenv
requests
